# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/quotes/{id}": {
            "get": {
                "tags": [
                    "Модель цитаты"
                ],
                "summary": "Получает цитату по ID",
                "description": "Получает цитату по ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID цитаты",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Отдает конкретную цитату по id"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}